Alert module fix for weird inconsistent state.
An issue where the alert was set to the Armed state but the external interrupt
was not being set. This was caused by a momentary invalid command of DISARM
being sent while the system was in OK state. This caused the state to change
without resetting the EXTINT register.

Also fix to the variable naming of the I2C frequency and addition of
status print on POR.